refactor work in progress
